Separate out all the actions that happen to the graph post-rendering.
This change breaks a few unit tests. It is a change in a sequence of several
changes to guarantee the order of asynchronous calls in graph rendering.
Subsequent changes will fix it, but the work was too big to make into
one PR, and there's no point in fixing the tests when at this point
their breakage is because the call ordering isn't guaranteed.

Additionally, I've reworked axis label wrapping here to be simpler.

Change-Id: Idf7c5906f67dfbb92dec00bbc07cd0c2b5dc6025
GitOrigin-RevId: 540b0c50a93ddba34a9ff5c81993b5d600b9dcb5