Clarify and simplify display range calculations for graphs.
The prior implemention had multiple places to specify "normal"
ranges. This made it unpredictable
and confusing to determine what should happen, and error-prone to code
as the same ranges had to be pasted multiple places.

This implementation makes it so that normal ranges are only associated
with LOINC codes and gives a clearer algorithm for which range will be
shown.

Change-Id: I43d8fee1cc931cd1cd4da42a655d1838a6094de3
GitOrigin-RevId: bf75b2701fe071b228f4f6587fca71ee0c91d6c6