Cache Results per day
For each of the resource groups (LOINCCodeGroup, RxNormGroup,
BCHMicrobioCodeGroup), cache the most basic results from FHIR on a day
by day basis. Then when we call for a new date range, check the cache
for days we have already gotten. Then request the date intervals that we
do not have in the cache already.

Because the results needed for rendering are not grouped by date, after
we fetch the data from the cache, we may need to format the results and
group in different ways.

For each subclass of resource code group, we need to create 2 functions:
1. getResourceFromFhir: this should get the raw results from FHIR. This
data will be put into the cache.
2. formatRawResults: this should format the raw results in the Cache
into the formatted (and grouped) results that the rendering code
expects.

Change-Id: I87939b04986cef5c5d8920348865e5acf8c7e81c
GitOrigin-RevId: 41e1ff344510a2d04f2a28ce5bd0eb394574753e